Refactor (gradio part 6) address CodeRabbit review findings
